Fleshed out main evaluate definition (still need to handle NFA states)
